Make the user able to hit the opponent's board
The opponent's board lights up red when the user misses an attack, and
lights up green when they hit an attack. A functionality will be added
that allows the robot to also makes moves on the user's board. Once
either the user or the opponent has all their ships sunk, the game
ends.